Getting the user now and populating the profile.
Reorganised the screens to use wrappers to contain the redux connect changes.

Need to look at the Organisations to see why the token is not set when trying to
use it in componentWillMount.

It might be that all the data should be got up front once the token exists rather
than trying to do it in the will mount method.